requirements.txt
Django==4.2
djangorestframework==3.14.0
isort==5.12.0
pre-commit==3.2.2
psycopg2==2.9.6
python-dotenv==1.0.0
